+++ cryptography/hazmat/backends/openssl/backend.py
@@ -742,10 +742,17 @@ class Backend(object):
         if not errors:
             raise ValueError("Could not unserialize key data.")
 
-        elif errors[0][1:] == (
-            self._lib.ERR_LIB_EVP,
-            self._lib.EVP_F_EVP_DECRYPTFINAL_EX,
-            self._lib.EVP_R_BAD_DECRYPT
+        elif errors[0][1:] in (
+            (
+                self._lib.ERR_LIB_EVP,
+                self._lib.EVP_F_EVP_DECRYPTFINAL_EX,
+                self._lib.EVP_R_BAD_DECRYPT
+            ),
+            (
+                self._lib.ERR_LIB_PKCS12,
+                self._lib.PKCS12_F_PKCS12_PBE_CRYPT,
+                self._lib.PKCS12_R_PKCS12_CIPHERFINAL_ERROR,
+            )
         ):
             raise ValueError("Bad decrypt. Incorrect password?")
 
@@ -977,6 +984,11 @@ class Backend(object):
         values.
         """
 
+        if x < 0 or y < 0:
+            raise ValueError(
+                "Invalid EC key. Both x and y must be non-negative."
+            )
+
         bn_x = self._int_to_bn(x)
         bn_y = self._int_to_bn(y)
 
@@ -998,16 +1010,16 @@ class Backend(object):
             res = get_func(group, point, check_x, check_y, bn_ctx)
             assert res == 1
 
-            assert (
-                self._lib.BN_cmp(bn_x, check_x) == 0 and
-                self._lib.BN_cmp(bn_y, check_y) == 0
-            )
+            assert self._lib.BN_cmp(bn_x, check_x) == 0
+            assert self._lib.BN_cmp(bn_y, check_y) == 0
 
         res = self._lib.EC_KEY_set_public_key(ctx, point)
         assert res == 1
 
         res = self._lib.EC_KEY_check_key(ctx)
-        assert res == 1
+        if res != 1:
+            self._consume_errors()
+            raise ValueError("Invalid EC key.")
 
         return ctx