+++ tests/conftest.py
@@ -16,13 +16,8 @@ from __future__ import absolute_import, division, print_function
 import pytest
 
 from cryptography.hazmat.backends import _available_backends
-from cryptography.hazmat.backends.interfaces import (
-    CMACBackend, CipherBackend, DSABackend, EllipticCurveBackend, HMACBackend,
-    HashBackend, PBKDF2HMACBackend, PEMSerializationBackend,
-    PKCS8SerializationBackend, RSABackend,
-    TraditionalOpenSSLSerializationBackend
-)
-from .utils import check_backend_support, check_for_iface, select_backends
+
+from .utils import check_backend_support, select_backends, skip_if_empty
 
 
 def pytest_generate_tests(metafunc):
@@ -30,26 +25,24 @@ def pytest_generate_tests(metafunc):
     selected_backends = select_backends(names, _available_backends())
 
     if "backend" in metafunc.fixturenames:
-        metafunc.parametrize("backend", selected_backends)
+        filtered_backends = []
+        for backend in selected_backends:
+            required = metafunc.function.requires_backend_interface
+            required_interfaces = tuple(
+                mark.kwargs["interface"] for mark in required
+            )
+            if isinstance(backend, required_interfaces):
+                filtered_backends.append(backend)
+
+        # If you pass an empty list to parametrize Bad Things(tm) happen
+        # as of pytest 2.6.4 when the test also has a parametrize decorator
+        skip_if_empty(filtered_backends, required_interfaces)
+
+        metafunc.parametrize("backend", filtered_backends)
 
 
 @pytest.mark.trylast
 def pytest_runtest_setup(item):
-    check_for_iface("hmac", HMACBackend, item)
-    check_for_iface("cipher", CipherBackend, item)
-    check_for_iface("cmac", CMACBackend, item)
-    check_for_iface("hash", HashBackend, item)
-    check_for_iface("pbkdf2hmac", PBKDF2HMACBackend, item)
-    check_for_iface("dsa", DSABackend, item)
-    check_for_iface("rsa", RSABackend, item)
-    check_for_iface(
-        "traditional_openssl_serialization",
-        TraditionalOpenSSLSerializationBackend,
-        item
-    )
-    check_for_iface("pkcs8_serialization", PKCS8SerializationBackend, item)
-    check_for_iface("elliptic", EllipticCurveBackend, item)
-    check_for_iface("pem_serialization", PEMSerializationBackend, item)
     check_backend_support(item)