+++ tests/hazmat/primitives/test_3des.py
@@ -22,6 +22,7 @@ import os
 
 import pytest
 
+from cryptography.hazmat.backends.interfaces import CipherBackend
 from cryptography.hazmat.primitives.ciphers import algorithms, modes
 
 from .utils import generate_encrypt_test
@@ -34,7 +35,7 @@ from ...utils import load_nist_vectors
     ),
     skip_message="Does not support TripleDES CBC",
 )
-@pytest.mark.cipher
+@pytest.mark.requires_backend_interface(interface=CipherBackend)
 class TestTripleDESModeCBC(object):
     test_KAT = generate_encrypt_test(
         load_nist_vectors,
@@ -71,7 +72,7 @@ class TestTripleDESModeCBC(object):
     ),
     skip_message="Does not support TripleDES OFB",
 )
-@pytest.mark.cipher
+@pytest.mark.requires_backend_interface(interface=CipherBackend)
 class TestTripleDESModeOFB(object):
     test_KAT = generate_encrypt_test(
         load_nist_vectors,
@@ -108,7 +109,7 @@ class TestTripleDESModeOFB(object):
     ),
     skip_message="Does not support TripleDES CFB",
 )
-@pytest.mark.cipher
+@pytest.mark.requires_backend_interface(interface=CipherBackend)
 class TestTripleDESModeCFB(object):
     test_KAT = generate_encrypt_test(
         load_nist_vectors,
@@ -145,7 +146,7 @@ class TestTripleDESModeCFB(object):
     ),
     skip_message="Does not support TripleDES CFB8",
 )
-@pytest.mark.cipher
+@pytest.mark.requires_backend_interface(interface=CipherBackend)
 class TestTripleDESModeCFB8(object):
     test_KAT = generate_encrypt_test(
         load_nist_vectors,
@@ -182,7 +183,7 @@ class TestTripleDESModeCFB8(object):
     ),
     skip_message="Does not support TripleDES ECB",
 )
-@pytest.mark.cipher
+@pytest.mark.requires_backend_interface(interface=CipherBackend)
 class TestTripleDESModeECB(object):
     test_KAT = generate_encrypt_test(
         load_nist_vectors,