+++ tests/hazmat/primitives/twofactor/test_totp.py
@@ -16,6 +16,7 @@ from __future__ import absolute_import, division, print_function
 import pytest
 
 from cryptography.exceptions import InvalidToken, _Reasons
+from cryptography.hazmat.backends.interfaces import HMACBackend
 from cryptography.hazmat.primitives import hashes
 from cryptography.hazmat.primitives.twofactor.totp import TOTP
 
@@ -27,7 +28,7 @@ vectors = load_vectors_from_file(
     "twofactor/rfc-6238.txt", load_nist_vectors)
 
 
-@pytest.mark.hmac
+@pytest.mark.requires_backend_interface(interface=HMACBackend)
 class TestTOTP(object):
     @pytest.mark.supported(
         only_if=lambda backend: backend.hmac_supported(hashes.SHA1()),